+++ lib/grax/rdf/loader.ex
@@ -110,9 +110,10 @@ defmodule Grax.RDF.Loader do
     map_values(objects, property_schema.type)
   end
 
-  defp map_values(values, {:list_set, _type}), do: map_while_ok(values, &map_value(&1))
-  defp map_values([value], _type), do: map_value(value)
-  defp map_values(values, _type), do: map_while_ok(values, &map_value(&1))
+  @doc !"This internal function is public because it is also used by Grax.Schema.Mapping"
+  def map_values(values, {:list_set, _type}), do: map_while_ok(values, &map_value(&1))
+  def map_values([value], _type), do: map_value(value)
+  def map_values(values, _type), do: map_while_ok(values, &map_value(&1))
 
   defp map_value(%Literal{} = literal) do
     if Literal.valid?(literal) do