+++ test/support/example_ns.ex
@@ -5,4 +5,9 @@ defmodule Example.NS do
     base_iri: "http://example.com/",
     terms: [],
     strict: false
+
+  defvocab FOAF,
+    base_iri: "http://xmlns.com/foaf/0.1/",
+    terms: [:Person, :foaf, :mbox],
+    strict: false
 end