+++ CHANGELOG.md
@@ -5,6 +5,17 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 [Keep a CHANGELOG](http://keepachangelog.com).
 
 
+## Unreleased
+
+### Added
+
+- Grax schema mapping functions `from/1` and `from!/1` on Grax schema modules,
+  which allows to map one schema struct to another
+
+[Compare v0.3.5...HEAD](https://github.com/rdf-elixir/grax/compare/v0.3.5...HEAD)
+
+
+
 ## v0.3.5 - 2023-01-18
 
 ### Fixed